Fixed Script und logonserver build. Fixed some compile warnings in scripts. We only need the icon in the .rc file and no file info, cleaned up logonserver.rc

+++ src/server/scripts/Instances/Icecrown_Citadel/Halls_of_Reflection/boss_falric.cpp
@@ -76,7 +76,7 @@ public:
 
        uint8 SummonCount;
 
-       uint64 pSummon;
+       uint32 pSummon;
 
         void Reset()
         {
